chore: move git and jscpd to lib folder
