+++ src/lib/jscpd.js
